Reorganize react component directories
