+++ package.json
@@ -57,6 +57,7 @@
     "typescript-plugin-css-modules": "3.4.0",
     "vite": "2.9.1",
     "vite-plugin-svgr": "1.1.0",
+    "vite-tsconfig-paths": "3.4.1",
     "wait-on": "6.0.1"
   },
   "engines": {
