+++ src/renderer/react/ReactApp.tsx
@@ -1,0 +1,8 @@
+import '../i18n'
+import WindowsRouter from './windows'
+
+export default function ReactApp (): JSX.Element {
+  return (
+    <WindowsRouter />
+  )
+}
