+++ src/renderer/react/components/MediaDisplayer/index.tsx
