+++ src/renderer/react/windows/ControlsWindow/FileList/FileItem/index.tsx
