+++ src/renderer/react/windows/ControlsWindow/FileList/Preview/index.module.css
