+++ src/renderer/react/windows/ControlsWindow/Menu/index.tsx
@@ -1,6 +1,6 @@
 import cx from 'clsx'
 import { useTranslation } from 'react-i18next'
-import { version } from '../../../../package.json'
+import { version } from '../../../../../../package.json'
 
 import MenuItem from './MenuItem'
 import styles from './index.module.css'
