+++ src/renderer/react/windows/ControlsWindow/VideoControls/ElapsedTimeBar/index.module.css
