+++ src/renderer/react/windows/ControlsWindow/VideoControls/index.module.css
