+++ src/renderer/react/windows/ControlsWindow/VideoControls/index.tsx
