+++ src/renderer/styles/colors.css
