+++ src/renderer/styles/global.css
