+++ vite.config.js
@@ -1,9 +1,10 @@
 import { defineConfig } from 'vite'
 import react from '@vitejs/plugin-react'
 import svgrPlugin from 'vite-plugin-svgr'
+import tsconfigPaths from 'vite-tsconfig-paths'
 
 export default defineConfig({
-  plugins: [react(), svgrPlugin()],
+  plugins: [react(), svgrPlugin(), tsconfigPaths({ root: '../../' })],
   root: './src/renderer',
   base: './',
   build: {
